Nutri-score/requirements.txt
pandas
python-csv
numpy
matplotlib.pyplot
seaborn
scikit-learn
missingno
imblearn
xgboost
lightgbm
catboost